+++ santashop-app/src/app/core/services/index.ts
@@ -1,4 +1,3 @@
-export * from './app-state.service';
 export * from './checkin.service';
 export * from './pre-registration.service';
 export * from './profile.service';
