+++ santashop-app/src/app/features/bad-weather/bad-weather.page.ts
@@ -1,11 +1,11 @@
 import { ChangeDetectionStrategy, Component, inject } from '@angular/core';
-import { AppStateService } from '../../core/services/app-state.service';
-import { CoreModule } from '@santashop/core';
+import { AppStateService, CoreModule } from '@santashop/core';
 import { AsyncPipe } from '@angular/common';
-import { TranslateModule } from '@ngx-translate/core';
+import { TranslateModule, TranslateService } from '@ngx-translate/core';
 import { addIcons } from 'ionicons';
 import { logoFacebook, logoInstagram } from 'ionicons/icons';
 import { IonContent, IonButton, IonIcon } from '@ionic/angular/standalone';
+import { map } from 'rxjs/operators';
 
 @Component({
 	selector: 'app-bad-weather',
@@ -23,8 +23,17 @@ import { IonContent, IonButton, IonIcon } from '@ionic/angular/standalone';
 })
 export class BadWeatherPage {
 	public readonly service = inject(AppStateService);
+	private readonly translateService = inject(TranslateService);
 
-	public readonly message$ = this.service.message$;
+	public readonly message$ = this.service.messageDoc$.pipe(
+		map((doc) => {
+			const message =
+				this.translateService.currentLang === 'en'
+					? doc.messageEn
+					: doc.messageEs;
+			return message?.length ? message : null;
+		}),
+	);
 
 	constructor() {
 		addIcons({ logoFacebook, logoInstagram });
