+++ santashop-app/src/app/features/v2/sign-in/sign-in.page.ts
@@ -22,35 +22,35 @@ import {
 } from '@ionic/angular/standalone';
 import { addIcons } from 'ionicons';
 import { arrowBackSharp } from 'ionicons/icons';
-import { AppStateService } from '../../../core';
+import { AppStateService } from '@santashop/core';
 import { AsyncPipe } from '@angular/common';
 
 @Component({
-    selector: 'app-sign-in',
-    templateUrl: './sign-in.page.html',
-    styleUrls: ['./sign-in.page.scss'],
-    changeDetection: ChangeDetectionStrategy.OnPush,
-    providers: [SignInPageService],
-    imports: [
-        AsyncPipe,
-        IonContent,
-        IonGrid,
-        IonRow,
-        IonCol,
-        IonCard,
-        IonCardHeader,
-        IonLabel,
-        IonCardTitle,
-        IonCardContent,
-        IonIcon,
-        IonList,
-        IonItem,
-        IonInput,
-        IonButton,
-        RouterLink,
-        ReactiveFormsModule,
-        TranslateModule,
-    ]
+	selector: 'app-sign-in',
+	templateUrl: './sign-in.page.html',
+	styleUrls: ['./sign-in.page.scss'],
+	changeDetection: ChangeDetectionStrategy.OnPush,
+	providers: [SignInPageService],
+	imports: [
+		AsyncPipe,
+		IonContent,
+		IonGrid,
+		IonRow,
+		IonCol,
+		IonCard,
+		IonCardHeader,
+		IonLabel,
+		IonCardTitle,
+		IonCardContent,
+		IonIcon,
+		IonList,
+		IonItem,
+		IonInput,
+		IonButton,
+		RouterLink,
+		ReactiveFormsModule,
+		TranslateModule,
+	],
 })
 export class SignInPage {
 	private readonly viewService = inject(SignInPageService);
