+++ santashop-app/src/environments/environment.prod.ts
@@ -1,4 +1,4 @@
-import config from '../../../package.json';
+import config from '../../package.json';
 
 export const environment = {
 	production: true,
